big config update
Config now lives as its own menu window. Changes to all of the settings
(with the exception of the new `dimension_units`, which I'll touch on
below) are now updated in the `app/local/settings.json` file and broadcast
to the app (where they're saved) and main window.

All that remains is adding a way to _get_ to the config (probably via
`Menu`) and setting up printing. Almost there!

Why no adjustable `dimension_units`?

Ideally, I'd like any adjustments to this field to update the values already
in place, and as we're only going to be using inches, I decided to put this
off until a later time.